Testing, QoL updates, vacation calculation

+++ src/dotnet/HQ.Abstractions/Times/GetDashboardTimeV1.cs
@@ -27,7 +27,7 @@ public class GetDashboardTimeV1
         public decimal HoursThisMonth { get; set; }
         public decimal HoursLastWeek { get; set; }
         public string? StaffName { get; set; }
-        public int Vacation { get; set; }
+        public decimal Vacation { get; set; }
 
     }
